// src/app/pages/login/reclutador/reclutador-interface.ts
export interface ReclutadorUpdate {
    nombreUsuario: string;
    ciudadUsuario: string;
    nombrecontactanteUsuario: string;
    tamañoempresaUsuario: string;
    telefonoUsuario: any;
    direccionUsuario: string;
    descripcionUsuario: string;
    tituloUsuario: string;  
}
export interface emailusuario{
    emailUsuario: string;
}